Re-add calendar commands
